lca now works
