+++ neosca/lca/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+# -*- coding=utf-8 -*-
+
+from .main import lca_main
+
+__all__ = ["lca_main"]
