+++ neosca/lca/__main__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding=utf-8 -*-
+
+from .main import lca_main
+
+if __name__ == "__main__":
+    lca_main()
